+++ taut.py
@@ -79,7 +79,7 @@ class dhaven:
                 
                 if getpotion:
                     if potreq > 20:
-                        for i in range(max([potreq/20-1,0])):
+                        for i in range(max([int(potreq/20)-1,0])):
                             self.rod.write("buy %s %s\nempty bag basket\ndrop bag\n"%(20, self.potname.split()[2]))
                     else:
                         self.rod.write("buy %s %s\nempty bag basket\ndrop bag\n"%(20, self.potname.split()[2]))
